[wip] gcc: generate nested function trampolines off stack
An initial attempt to support nested functions on Darwin/AArch64,
where the stack is marked non-execute.

Nested function trampolines are allocated into off-stack pages, which
are marked as executable.  The trampolines are deallocated using a
finally block in the scope of the function containing the nested
function.

The only problem with this is that nested function trampolines will be
leaked if a user longjmps over the deallocation code (as the finally
block will not be executed).  We could potentially look to solve this
problem by hooking into the setjmp/longjmp code within GCC, and
trying to understand which trampolines have been jumped over (and so
need releasing), however, for now I'm considering that a separate,
second stage of the project.

The current code is very much work in progress, there's a mountain of
things that still need doing:

(1) The choice between on-stack and off-stack trampolines is currently
controlled by a global in tree-nested.c, this is obviously just a
temporary back, and needs a real control option, but...

(2) ... I suspect some of the on-stack trampoline code may have been
broken by my hacking in the off-stack trampoline support, so we need
to audit all of the code changes and make sure that I've not broken
the traditional way of doing things.

(3) The current library code to allocate and deallocate trampolines
needs to be moved to a new home, it probably also needs to be
configure-time selectable, as targets that have an executable stack
are unlikely to want to use the off-stack approach (at least for now).

(4) The library code needs to be updated in two respects:

    (a) we need to support generating more than a single page of
    trampolines, this will require us to be able to allocate
    additional pages and chain them together, then of course, we
    should possibly be deallocating pages as trampolines are released,
    and

    (b) we should generate trampolines by copying in some template
    code, and then patching it with the appropriate addresses, this is
    not what we currently do.