Implement INIT_CUMULATIVE_INCOMING_ARGS for AArch64 Darwin
In order to have a matching ABI interface, both the callee and the
caller invocation of `locate_and_pad_parm` need to know which argument
is the last named argument. This is naturally faciliatated for the
caller expansion, but for the callee expansion the
INIT_CUMULATIVE_ARGS hook is called with `n_named_args` set to -1.

GCC permits us to optionally override the CUMULATIVE_ARGS
initialization for this case, which we can use to preserve the number
of named arguments we're called with and hence correctly access stack
parameters.