libstdc++: Adjust build of PCH files accounting configured host support.
This takes account of the overall configuration for host PCH support
when deciding if we should build the libstdc++ PCH files.

We now require both the support is configured and that we are hosted.
A non-fatal configure warning is given if the user attempts to
--disable-host-pch-support --enable-libstdcxx-pch since the latter
conflicts with the former (but does not prevent a useable libstdc++
library build).

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>

libstdc++-v3/ChangeLog:

	* acinclude.m4: Account for configured host PCH support.
	* configure: Regenerate.
	* configure.ac: Act on --enable-host-pch-support.